% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadContainers.R
\name{downloadContainers}
\alias{downloadContainers}
\title{Download for the first time all containers embedded in the workflows}
\usage{
downloadContainers(containers.file = NULL, tag = "latest")
}
\arguments{
\item{containers.file, }{a character string with the name of the file which indicate which are the initial set of containers to be downloaded. If NULL then the set is given by a file called "containersNames.txt" located in the folder inst/Containers of epimod package.}
}
\description{
This is a function that prepares the docker environment to be used for the first time the epimod is installed.
}
\examples{
\dontrun{
    #running runDocker
     downloadContainers()

}
}
\author{
Beccuti Marco, Castagno Paolo, Pernice Simone
}
