% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.event.cmd.R
\name{event.worker}
\alias{event.worker}
\title{Generate the commandline to configure and run a simulation}
\usage{
event.worker(id, solver_fname, solver_type, s_time, f_time, n_run = 1,
  taueps, timeout, out_fname, times.events, marking.delta)
}
\arguments{
\item{id, }{a numeric identifier used to format the output's file name}

\item{solver_fname, }{the name of the solver executable file}

\item{solver_type, }{a string definig what solver to apply (LSODA, HLSODA, ..)}

\item{s_time, }{step time at which the sover is forced to output the current configuration of the model (i.e. the number of tocken in each place)}

\item{f_time, }{simulation's final time}

\item{n_run, }{when performing stochastic simulations this parameters controls the number of runs per each set of input parameters}

\item{timeout, }{string controlling the available time to run n_run simulations. See TIMEOUT(1) to check the syntax}

\item{out_fname, }{output filename prefix}
}
\value{
a string
}
\description{
This is an internal function generating the command line to run the solver with the appropriate configuration
}
\examples{
\dontrun{
experiment.cmd(id=1, solver_fname="Solver.solver", s_time=365, f_time=730, timeout="1d", out_fname="simulation")

}
Example of events list:
event.list<-list(e1=list(time = 4 , update = c(+1,0,0,-1)),
                 e2=list(time = 8 , update = c(+2,0,0,-1)))

}
\author{
Paolo Castagno
}
