% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.cmd.R
\name{experiment.cmd}
\alias{experiment.cmd}
\title{Generate the command template to configure and run a simulation}
\usage{
experiment.cmd(
  solver_fname,
  solver_type = "LSODA",
  taueps = 0.01,
  seed,
  timeout
)
}
\arguments{
\item{solver_fname, }{the solver executable file}

\item{solver_type, }{the type of solver (LSODA, HLSODA, ..)}

\item{taueps, }{controls the time step for tau-leap algorithm}

\item{timeout, }{available time to run n_run simulations (See TIMEOUT(1) to check the syntax)}

\item{event_times, }{times at which the marking is updated through the event_function}

\item{event_function, }{specifies the rule to update the marking}

\item{out_fname, }{output filename prefix}
}
\value{
a string
}
\description{
Function generating the command line used to run the solver with the appropriate configuration
}
\examples{
\dontrun{
experiment.cmd(id=1, solver_fname="Solver.solver", s_time=365, f_time=730, timeout="1d", out_fname="simulation")

}
}
\author{
Paolo Castagno, Simone Pernice
}
