% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_configuration.R
\name{experiment.configurations}
\alias{experiment.configurations}
\title{Generates a set of simulation's configurations}
\usage{
experiment.configurations(n_config, parm_fname, parm_list, out_dir,
  out_fname, extend = "", optim_vector = NULL)
}
\arguments{
\item{n_config, }{number of configuratons to generate}

\item{parm_fname, }{file with the definition of user defined functions}

\item{parm_list, }{file listing the name of the functions, the parameters and the name under which the parameters have to be saved}

\item{out_dir, }{output directory specified by the user}
}
\description{
This is an internal function generating a set of configurations. The generation process is driven by the functions listed in param_list. Such functions either are native R functions or user defined ones. In the latter case, the functions have to be defined in a single file and passed through the argument parm_fname
}
\examples{
\dontrun{
experiment.env_cleanup(id=1, run_dir="/run/directory", out_fname="simulation", out_dir="/out/directory",)
}
}
\author{
Paolo Castagno
}
