% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.env_cleanup.R
\name{experiment.env_cleanup}
\alias{experiment.env_cleanup}
\title{Removes unnexessary files and directories}
\usage{
experiment.env_cleanup(id, run_dir, out_fname, out_dir)
}
\arguments{
\item{id, }{a numeric identifier used to format the output's file name}

\item{run_dir, }{the directory to cleanup}

\item{out_fname, }{output filename prefix}

\item{out_dir, }{output directory specified by the user}
}
\description{
This is an internal function removing unnecessary files (and the running directory itself) from the running environment
}
\examples{
\dontrun{
experiment.env_cleanup(id=1, run_dir="/run/directory", out_fname="simulation", out_dir="/out/directory",)
}
}
\author{
Paolo Castagno
}
