% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.env_setup.R
\name{experiment.env_setup}
\alias{experiment.env_setup}
\title{Sets up the environment (directory and files) required to run a simulartion}
\usage{
experiment.env_setup(id = NULL, files, dest_dir, config = NULL)
}
\arguments{
\item{id, }{a numeric identifier used to format the output's file name (optional, if not provided copy files in dest_dir. Otherwise, create a sub-directory id)}

\item{confg, }{configuration provided by the function experiment.configurations}
}
\description{
This is an internal function creating the required structure of files and directiry to run a simulation.
}
\examples{
\dontrun{
experiment.env_cleanup(id=1, run_dir="/run/directory", out_fname="simulation", out_dir="/out/directory",)
}
}
\author{
Beccuti Marco, Castagno Paolo, Pernice Simone
}
\keyword{internal}
