% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_calibration.R
\name{model_calibration}
\alias{model_calibration}
\title{Run model calibration}
\usage{
model_calibration(out_dir, out_fname, parameters_fname = "",
  functions_fname = "", solver_fname, solver_type = "LSODA",
  init_fname = NULL, f_time, s_time, n_run = 1, volume = "",
  timeout = "1d", processors, ini_v, lb_v, ub_v, nb.stop.improvement,
  reference_data, distance_measure_fname, distance_measure,
  target_value_fname, target_value_f, seed = NULL, extend = NULL)
}
\arguments{
\item{out_dir}{Directory where to store all the output generated}

\item{out_fname}{Prefix to the output file name}

\item{functions_fname}{File with the user defined functions to generate istances of the parameters}

\item{solver_fname}{.solver file (generated in with the function model_generation)}

\item{f_time}{Final solution time}

\item{s_time}{Time step at whicch explicit estimates for the system are desired}

\item{volume}{The folder to mount within the Doker image providing all the necessary files}

\item{reference_data}{Data to compare with the simulations' results}

\item{distance_measure_fname}{File containing the definition of a distance measure to rank the simulations'. Such function takes 2 arguments: the reference data and a list of data_frames containing simulations' output. It has to return a data.frame with the id of the simulation and its corresponding distance from the reference data}

\item{distance_measure}{The name of the function defining the distance measure}

\item{config_fname}{File collecting all the configurations generated in the sensitivity analysis}

\item{int_fname}{File containing the initial marking, if required}
}
\description{
this functon takes as input a solver and all the required parameters to set up a dockerized running environment to perform model calibration. both for deterministic and stochastic models
In order to run the simulations, the user must provide a reference dataset and the definition of a function to compute the distance (or error) between the models' output and the reference dataset itself.
The function defining the distance has to be in the following form:

FUNCTION_NAME(reference_dataset, siulation_output)

Moreover, the function must return a column vector with one entry for each evaluation point (i.e. f_time/s_time entries)
in addiction to that, the user is asked to provide a function that, given the output of the solver, returns the releveant measure (one column) used to evalaute the quality of the solution.

The sensitivity analysis will be performed through a Monte Carlo sampling throug user defined functions.
the parameters involved in the sensitivity analysis have to be listed in a cvs file using the following structure:

OUTPUT_FILE_NAME, FUNCTION_NAME, LIST OF PARAMETERS (comma separated)

The functions allowed to compute the parameters are either R functions or user defined functions. In the latter case, all the user defined functions must be provided in a single .R file (which will be passed to run_sensitivity through the parameter parameters_fname)

Exploiting the same mechanism, user can provide an initial marking to the solver. However, if it is the case the corresponding file name in the parameter list must be set to "init"

To drive the optimization, the user has to provide a function to generate a new configuration, starting from a vector of n elements (each one ranging from 0 to 1).
Furthermore, the vector init_v defines the initial point of the search.

IMPORTANT: the length of the vector init_v defines the number of variables to variate within the search of the optimal configuration.
}
\author{
Paolo Castagno
}
