% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_generation.R
\name{model_generation}
\alias{model_generation}
\title{Run model generation}
\usage{
model_generation(
  net_fname,
  out_fname = NULL,
  transitions_fname = NULL,
  volume = getwd()
)
}
\arguments{
\item{net_fname}{.PNPRO file storing  the Petri Net (and all its generalizations) model. In case there are multiple nets defined within the PNPRO file, the first one in the list is the will be automatically selected.}

\item{out_fname}{Prefix to the output file name.}

\item{transitions_fname}{C++ file defining the functions managing the behaviour of general transitions, mandatory if Extended versions of Petri Nets (i.e., ESPN or ESSN) are used.}

\item{volume}{The folder to mount within the Doker image providing all the necessary files.}
}
\value{
Model_generation returns the binary file SIR.solver in which the underlying processes (both deterministic and stochastic) of the Petri Net model and the library used for their simulation are packaged.
}
\description{
Having constructed the model exploiting the graphical editor, namely GreatSPN, the automatic generation of both the stochastic (the Continuous Time Markov Chain) and
     deterministic (ODEs) processes underlying the model is implemented by the model_generation() function.
}
\details{
GreatSPN GUI, the graphical editor for drawing Petri Nets formalisms, is available online: http://www.di.unito.it/~amparore/mc4cslta/editor.html
}
\examples{
\dontrun{
local_dir <- "/some/path/to/the/directory/hosting/the/input/files/"
model_generation(net_fname = paste0(local_dir, "Configuration/Pertussis"),
                 out_fname = "Solver",
                 transitions_fname = "transitions.cpp")
}

}
\author{
Beccuti Marco, Castagno Paolo, Pernice Simone
}
