% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_check.R
\name{results_check}
\alias{results_check}
\title{Check the results files}
\usage{
results_check(
  fname_st = NULL,
  fname_nd = NULL,
  furl_st = NULL,
  fun,
  threshold = NULL
)
}
\arguments{
\item{fname_st}{Path to the first file to compare}

\item{fname_nd}{Path to the second file to compare}

\item{furl_st}{Url of the first file to compare}

\item{fun}{A string that defines which control algorithm apply (det_check, sto_check, fda_test)}

\item{threshold}{A threshold used to compare the p-values generated in functional data analysis}
}
\description{
This function can compare the results obtained by a deterministic model, in this case
the files are compared using the diff unix command, obtained by two different stochastic model
calculating the confidence intervals and do a simple form of functional data analysis using
the fdatest package. Produce a file log with the output of the used functions.
}
\author{
Daniele Baccega, Luca Rosso
}
