% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispaly_data.R
\name{stop_display}
\alias{stop_display}
\title{Stop the display container}
\usage{
stop_display(silent = FALSE)
}
\arguments{
\item{silent}{If TRUE suppresses messages when no container is running. Defaults to FALSE.}
}
\description{
Stops and removes the running EpiMod Plot Viewer Docker container.
}
\examples{
\dontrun{
stop_display()
}
}
\author{
Beccuti Marco, Castagno Paolo, Pernice Simone
}
