% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispaly_data.R
\name{display_data}
\alias{display_data}
\title{Run display data}
\usage{
display_data(volume = getwd(), port = 80)
}
\arguments{
\item{volume}{The folder to mount within the Doker image providing all the necessary files}

\item{port}{The port shyiny app will listen for requests}
}
\description{
Run display data
}
\examples{
\dontrun{
display_data(volume = "/some/path/to/the/local/output/directory",
              port = getOption("shiny.port"))
}
}
\author{
Beccuti Marco, Castagno Paolo, Pernice Simone
}
