% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.run.R
\name{experiment.run}
\alias{experiment.run}
\title{Run an experiment, handling the eventual discrete events}
\usage{
experiment.run(
  id,
  cmd,
  i_time,
  f_time,
  s_time,
  atol,
  rtol,
  achn,
  rchn,
  n_run = 1,
  seed,
  event_times = NULL,
  event_function = NULL,
  out_fname,
  FVA = F
)
}
\arguments{
\item{id}{a numeric identifier used to format the output's file name}

\item{cmd}{the command line string to run the solver}

\item{i_time}{initial solution time}

\item{f_time}{simulation's final time}

\item{s_time}{step time at which the solver is forced to output the current configuration of the model}

\item{atol}{Absolute error tolerance that determine the error control performed by the LSODA solver.}

\item{rtol}{Relative error tolerance that determine the error control performed by the LSODA solver.}

\item{achn}{Absolute change tolerance for triggering an FBA update (absolute value).}

\item{rchn}{Relative change tolerance for triggering an FBA update (percentage value, e.g., 1%, 50%, 0.001%).}

\item{n_run}{when performing stochastic simulations this parameter controls the number of runs per each set of input parameters}

\item{seed}{.RData file that can be used to initialize the internal random generator.}

\item{event_times}{controls the time at which the simulation is stopped to update the marking}

\item{event_function}{specifies the rule to update the marking}

\item{out_fname}{output filename prefix}

\item{FVA}{Flag to enable the flux variability analysis}
}
\value{
the name of the trace file
}
\description{
This is an internal function generating the command line to run the solver with the appropriate configuration
}
\examples{
\dontrun{
experiment.cmd(id=1, solver_fname="Solver.solver", s_time=365, f_time=730, timeout="1d", out_fname="simulation")

}
}
\author{
Paolo Castagno, Simone Pernice
}
