% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.generation.R
\name{model.generation}
\alias{model.generation}
\title{Run model generation}
\usage{
model.generation(
  out_fname = NULL,
  net_fname,
  transitions_fname = NULL,
  fba_fname = NULL,
  volume = getwd(),
  debug = FALSE
)
}
\arguments{
\item{out_fname}{Prefix to the output file name.}

\item{net_fname}{.PNPRO file storing the Petri Net (and all its generalizations) model. In case there are multiple nets defined within the PNPRO file, the first one in the list is the will be automatically selected.}

\item{transitions_fname}{C++ file defining the functions managing the behaviour of general transitions, mandatory if Extended versions of Petri Nets (i.e., ESPN or ESSN) are used.}

\item{fba_fname}{vector of .txt files encoding different flux balance analysis problems, which as to be included in the general transitions (*transitions_fname*). (default is NULL)}

\item{volume}{The folder to mount within the Docker image providing all the necessary files.}

\item{debug}{If TRUE enables logging activity.}
}
\value{
model.generation returns the binary file SIR.solver in which the underlying processes (both deterministic and stochastic) of the Petri Net model and the library used for their simulation are packaged.
}
\description{
Having constructed the model exploiting the graphical editor, namely GreatSPN, the automatic generation of both the stochastic (the Continuous Time Markov Chain) and
  deterministic (ODEs) processes underlying the model is implemented by the model_generation() function.
}
\details{
GreatSPN GUI, the graphical editor for drawing Petri Nets formalism, is available online: http://www.di.unito.it/~amparore/mc4cslta/editor.html
}
\examples{
\dontrun{
local_dir <- "/some/path/to/the/directory/hosting/the/input/files/"
model.generation(out_fname = "Solver",
                 net_fname = paste0(local_dir, "Configuration/Pertussis"),
                 transitions_fname = "transitions.cpp")
}

}
\author{
Beccuti Marco, Castagno Paolo, Pernice Simone, Baccega Daniele
}
